python-server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
praw>=7.0.0
sentence-transformers>=2.0.0
umap-learn>=0.5.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
python-multipart>=0.0.5
